split dataset-create test
